+++ src/Kiota.Builder/Configuration/GenerationConfiguration.cs
@@ -105,14 +105,14 @@ public class GenerationConfiguration : ICloneable
     {
         get; set;
     }
-    public HashSet<string> StructuredMimeTypes
+    public StructuredMimeTypesCollection StructuredMimeTypes
     {
         get; set;
-    } = new(4, StringComparer.OrdinalIgnoreCase) {
-        "application/json",
-        "text/plain",
-        "application/x-www-form-urlencoded",
-        "multipart/form-data",
+    } = new StructuredMimeTypesCollection {
+        "application/json;q=1",
+        "text/plain;q=0.9",
+        "application/x-www-form-urlencoded;q=0.2",
+        "multipart/form-data;q=0.1",
     };
     public HashSet<string> IncludePatterns { get; set; } = new(0, StringComparer.OrdinalIgnoreCase);
     public HashSet<string> ExcludePatterns { get; set; } = new(0, StringComparer.OrdinalIgnoreCase);
@@ -139,7 +139,7 @@ public class GenerationConfiguration : ICloneable
             Serializers = new(Serializers ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
             Deserializers = new(Deserializers ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
             CleanOutput = CleanOutput,
-            StructuredMimeTypes = new(StructuredMimeTypes ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
+            StructuredMimeTypes = new(StructuredMimeTypes ?? Enumerable.Empty<string>()),
             IncludePatterns = new(IncludePatterns ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
             ExcludePatterns = new(ExcludePatterns ?? Enumerable.Empty<string>(), StringComparer.OrdinalIgnoreCase),
             ClearCache = ClearCache,
@@ -164,7 +164,7 @@ public class GenerationConfiguration : ICloneable
         if (languageInfo.StructuredMimeTypes.Any() &&
             comparer.Equals(StructuredMimeTypes, defaultConfiguration.StructuredMimeTypes) &&
             !comparer.Equals(languageInfo.StructuredMimeTypes, StructuredMimeTypes))
-            StructuredMimeTypes = new(languageInfo.StructuredMimeTypes, StringComparer.OrdinalIgnoreCase);
+            StructuredMimeTypes = new(languageInfo.StructuredMimeTypes);
     }
 }
 #pragma warning restore CA1056